Support for lax.js through Lax component and service
There's a Lax component and a Lax service, both of which can be used.
When you use the Lax component you basically have no setup, given the
component doesn't add contents.  An update action is triggered in case
you'd update the contents and need ot trigger Lax to update its
contents.  The service can be used to trigger lax's updates if you
don't want to use the component.